Bugfix: fix timefilter error for parallel script
problem
=======
for timefilter, it needs to fast forward to the buffer with satified time range,
pt will lookahead each buffer for each pt queue, but not add buffer to decoder.
The buffer will be added later. However, for parallel script, the lookahead
does add the buffer.

Solution
=======
add buf_lookahead and last_psb_lookhead to mark the lookhead progress, but not
add to decoder.
we also redirect intel pt debug message to different files for child process.