[feature]: support setting result dir to save/analyze trace data
Sometimes, we want to compare the latency of multiple trace data,
to find the status difference of the program. Currently we need to
switch directory by ourself.

In this patch, we add one option "-D/--result_dir" to set the result
directory to save/use the trace data. we just modify the result
directory, and we can see latency analysis of different trace data.