[feature] add irq latency, resolve_path and replace dso path for perf.data
1. Show the latency of interruption (apic timer, nmi, irq), which can also
be treated as child function. This will be enabled only in off-cpu mode,
and the call-line is set to "[gathered]" for simplicity, because it can be
called from all locations randomly when the cpu is high. We can show the
actual call-line of interruption by set "-U" to upfolding.

2. By default, when we do perf record, the executed binary path is persisted
to the mmap2 event of "perf.data", and we need to guarantee that the binary in
this path has not change, but sometimes it is difficult. Now we replace the dso path
in thread map when we do perf script if the "-b" is set and the execution binary
name is equal to our given dso name. Note it may also replace some lib dso if the
lib dso name is equal to the binary name.

3. resolve binary path with "~" and relative path.

4. optimize the script performance by remove enter perf_sample construction function
in intel_pt_synth_branch_sample.

5. Implement the subcommand like "./func_latency -b [binary] -f [func_name] -- [subcommand]",
but now for intel_pt, it doesn't trace the forked child thread.